+++ Commons/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## UNRELEASED
+### Added
+ - OperationResponseMessage and OperationResponseMessageType.
+### Changed
+ - Renamed Report to Message on OperationResponseBase.
+ - Renamed Retified to Rectified on SurfaceFinish.
+ - Renamed class OperationResponseBase to OperationResponse.
+
 ## [1.5.0] - 2022-02-24
 ### Added
  - Class BasePath.